A new look for your Makefile.PL
Hi Audrey!!

A present for you...

I just took some time to convert your setup to Module::Package.

I had to put out a new M:P version with a Unicode tweak for the Chinese package name.

I also had to alter your code slightly to account for missing code in M:I.

So consider this a rough draft. I hope you release it, and then work with me to make it better, a step at a time.

Cheers! Ingy